pialeda(main): bug fixed

+++ src/main/java/pialeda/app/Invoice/repository/ClientRepository.java
@@ -8,6 +8,6 @@ import pialeda.app.Invoice.model.Client;
 @Repository
 public interface ClientRepository extends JpaRepository<Client, Integer>{
     Client findByName(String name);
-    Client findById(long clientId);
-    Client deleteById(long id);
+    Client findById(int clientId);
+    Client deleteById(int id);
 }
